restructure plugins folder
