+++ file_compare/__init__.py
@@ -1,6 +1,6 @@
 from .module import ComparisonController
-from .plugins.videoPlugin import VideoPlugin
-from .plugins.imagePlugin import ImagePlugin
+from .plugins.video import VideoPlugin
+from .plugins.image import ImagePlugin
 
 from .base.compUtils import EnumGet
 from .base.compPlugin import ComparisonPlugin
