+++ file_compare/plugins/common/hasher.py
