+++ file_compare/plugins/image.py
@@ -1,6 +1,6 @@
 from PIL import Image
 from pillow_heif import register_heif_opener
-from .hasher import Hasher
+from .common.hasher import Hasher
 from file_compare.base.compPlugin import ComparisonPlugin, EnumGet
 from file_compare.base.compFilePlugin import FileAlgos
