+++ tests/test_plugins.py
@@ -67,8 +67,8 @@ def _get_test_files(plugin: type[ComparisonPlugin], files: list[Path]):
 
 if __name__ == "__main__":
     from file_compare.base.compFilePlugin import FilePlugin
-    from file_compare.plugins.imagePlugin import ImagePlugin
-    from file_compare.plugins.videoPlugin import VideoPlugin
+    from file_compare.plugins.image import ImagePlugin
+    from file_compare.plugins.video import VideoPlugin
 
     validate_plugins(
         THIS_PATH.parent,
